add some basic automated tests for isstudent + gh actions
